Inline context window display toggle on token usage element
Click the token usage display (circle or text) to toggle between
compact and detailed modes via PopoverMenu. Removes the separate
ContextWindowDisplaySelector component. Restores upstream footer
layout order.
